everything should be there now... except for the main application logic. wonder if i should put that into main.. tests would be nice, too, but i guess i will open that cmake hell later on

+++ IR-Arduino/IR-Arduino.ino
@@ -7,9 +7,9 @@ void setup() {
   tempSensor.begin();
 }
 
-double getTemperature() {
+float getTemperature() {
   // readObjectAmbient? see adafruit...
-  return tempSensor.readObjectTempC();
+  return (float) tempSensor.readObjectTempC();
 }
 
 void loop() {
